# frontend/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermaOps local API",
    "version": "2.0.0",
    "description": "Historical decision support. No hardware actuation. All numerical inputs must be finite."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/config": {
      "get": {
        "summary": "Sites and algorithms",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/states": {
      "get": {
        "summary": "Paginated historical states",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "dc",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DC_1",
                "DC_2",
                "DC_3",
                "DC_4",
                "DC_5"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000
            }
          }
        ],
        "description": "Returns baseline_feasible (all saved checks), baseline_region_valid, region_reachable (necessary but not sufficient for feasibility), minimum_region_distance, policy_revision and support_summary. Region reachability is the minimum continuous Mahalanobis distance over saved control bounds at fixed observed state."
      }
    },
    "/api/states/{state_id}": {
      "get": {
        "summary": "Historical state, baseline and simulation",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "state_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Includes region_support with minimum_distance, threshold, reachable and largest_state_shifts relative to training."
      }
    },
    "/api/trends": {
      "get": {
        "summary": "Recorded trends",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "dc",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DC_1",
                "DC_2",
                "DC_3",
                "DC_4",
                "DC_5"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 2,
              "maximum": 300
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "State ID belonging to the selected site"
          }
        ]
      }
    },
    "/api/limits/{dc}": {
      "get": {
        "summary": "Read site limits",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "dc",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "Save limits using optimistic concurrency",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "JSON error with a human-readable error field",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "dc",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "expected_revision",
                  "limits"
                ],
                "properties": {
                  "expected_revision": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "limits": {
                    "$ref": "#/components/schemas/Limits"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Limits": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "max_inlet_c",
          "max_hotspot_c",
          "reserve_c",
          "max_power_kw",
          "bounds"
        ],
        "properties": {
          "max_inlet_c": {
            "type": "number",
            "minimum": 18,
            "maximum": 27
          },
          "max_hotspot_c": {
            "type": "number",
            "minimum": 20,
            "maximum": 35
          },
          "reserve_c": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "max_power_kw": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 1,
            "maximum": 10000
          },
          "bounds": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "cooling_setpoint_c",
              "supply_air_temperature_c",
              "fan_speed_pct",
              "pump_speed_pct"
            ],
            "properties": {
              "cooling_setpoint_c": {
                "type": "array",
                "minItems": 2,
                "maxItems": 2,
                "items": {
                  "type": "number",
                  "multipleOf": 0.1,
                  "minimum": 18,
                  "maximum": 27
                }
              },
              "supply_air_temperature_c": {
                "type": "array",
                "minItems": 2,
                "maxItems": 2,
                "items": {
                  "type": "number",
                  "multipleOf": 0.1,
                  "minimum": 18,
                  "maximum": 27
                }
              },
              "fan_speed_pct": {
                "type": "array",
                "minItems": 2,
                "maxItems": 2,
                "items": {
                  "type": "number",
                  "multipleOf": 0.1,
                  "minimum": 30,
                  "maximum": 100
                }
              },
              "pump_speed_pct": {
                "type": "array",
                "minItems": 2,
                "maxItems": 2,
                "items": {
                  "type": "number",
                  "multipleOf": 0.1,
                  "minimum": 30,
                  "maximum": 100
                }
              }
            }
          }
        }
      }
    }
  }
}